Break up ANSI values
In other words, RED went from this:

val RED = "\u001B[31m"

to this:

object RED : SgrCode("31m")

At this point, we can process escape codes in pieces, instead of doing
full string comparisons.

We are not taking advantage of this yet in SwingTerminal, which
currently has logic that iterates through all keys in a table, but we
could eventually.