The great refactor
Now that we're getting close to MVP, I took some time to think about the
API layout, and I also used Compose for some inspiration. At the moment,
we have this:

- runtime
  All core bits that power the Konsole framework. The package should be
  self-sufficient, not depending on other packages.

- foundation
  Bits built on top of core to make the end-user API that feels good to
  use. Tons of extension methods in here, and some other useful classes.

- terminal
  Terminal implementations (of the runtime.terminal.Terminal interface)

Terminal can depend on runtime
Foundation can depend on terminal and runtime

After this change, there's still one dependency on foundation from
runtime (related to reading keypresses), but we'll save that for a
followup change because it might be a bit more involved.

Bug #23