Remove obsolete "tested" section from README
As a single dev, it's probably better to leave this out for now. I'm
sure if people use this library, they'll start reporting bugs for their
systems if it's not working for them.

For what it's worth, I've been checking this often on linux and
occasionally on Mac and it seems to be working in those places at least,
and Windows currently force defaults to the Virtual Terminal.