Remove InputLifecycle
Instead, input is tied to the section lifecycle and has its resources
suspended, instead of disposed, when it isn't being called.

This is done so we can keep track of input state even if input() is
not run for a few frames. I need this in a different program I'm
working on where I can tab away from an editor panel but still want
to come back to it later and have my cursor position remembered.