Change `run` logic so that exceptions are thrown on the main thread
I overcomplicated run's design originally. I had it run on a
background thread but then blocked the main thread until that
background thread had finished.

This made the use of Kotter feel right (run blocked the main
thread) but as a side effect, when run threw an exception,
it was lost.

Now, we use `runBlocking` to get the same behavior but less logic,
plus exceptions will now propogate to the user as expected.

(Technically, we still catch the exception ourselves before
rethrowing it, to allow us to do a small amount of state cleanup
when we detect the problem).

Fixes #71