Nit: README tweaks for readability
I did a runthrough of the README this morning and found a few areas
were hard for me to follow. Since I wrote it in the first place and
knew what I was trying to say, I thought I'd clean those areas up a
little. It's already better, I think, but I'll do another pass with
fresh eyes later.