Add "onPreRender" to section
It is useful for any test that might take several actions which,
behind the scenes, trigger multiple rerender requests.

We use it, as well, in TextAnimTest, which in some cases spawns
two animations at the same time (where each makes a separate
rerender request). By blocking until we're ready (when both anims
have finished making a render request), we can guarantee
that only one rerender will happen consistently, which we need
to be able to write consistent test assertions.