Move anim away from inheriting CharSequence
It seemed like a clever idea at the time but it adds extra fields
to the API that are kind of confusing, e.g. "length" which is
different from "numFrames".

The reason I made the class inherit CharSequence in the first place
was so that I could pass it into text and textLine. Instead, I just
create two new text and textLine extension methods that take a
textAnim as an argument.