Fix K/N issue with providing TIOCGWINSZ
The common posix layer for Linux and Mac diverges enough so
that it doesn't provide access to the TIOCGWINSZ constant,
despite it looking the same at a glance (both using the
symbol platform.posix.TIOCGWINSZ).

We work around this limitation by asking the Linux and Mac
layers separately to provide the symbol.