Stop applying render state for empty strings
In more technical terms, make "RenderScope.text(...)" aware of the
case where it could receive the empty string as input, and, if it
does, just abort early.

This isn't for some performance optimization. Rather, before, empty
text commands could get fed into the pipeline, causing Kotter to
force apply some styles prematurely. No need to do that if there's
no text!

In practice, this would probably happen very very rarely, and it
would not have been the biggest issue when it didn happen, but this
was making one test act very strangely, so it's nice to have figured
out the root cause.

Fixes #90