Introduce MockTerminal
... and refactor test code to use it.

It turns out it can be useful to expose a concept of an
in-memory terminal even for public users. For example,
I have another project that needs to implement an
interface that wants me to specify ASCII values directly,
and at the moment, there's no such way to pull them out of
Kotter.