Opt into posix APIs so K/N can compile again
This might actually be discouraged (see also
https://slack-chats.kotlinlang.org/t/15890271/anyone-else-adding-file-optin-experimentalforeignapi-class-t)
but in practice I don't think we have that many
K/N users right now, and I don't think we're using
anything that's not super fundamental in the posix
APIs at the moment.

It's possible that Kotter may break in the future
for native targets if JetBrains massively changes
their own approach. But if that would happen, it
would probably be such a big change that the user's
program would also need major work. So I'm going to
sidestep that issue for now.

We can revisit this if users start filing bugs.