Tie many input keys to the session lifecycle
... instead of the section lifecycle.

This is unfortunate but eliminates a very, very difficult timing
bug that occurs because the base flow (which converts system input
into Kotter `Key` instances) is always running while the consuming
flows are started and stopped per section, which due to factors out
of our control, can result with their handlers taking milliseconds to
connect, at which point any test that creates multiple sections that
all have inputs in them have a chance to fail.

Ideally we want to make all the input jobs tied to the Section
lifecycle (so if a section doesn't do any work with input, then
no resources will be used), but this will still have the tricky
timing issues mentioned above, so it may not be trivial.

Fixes #114