Get all input jobs tied to the section lifecycle
This now means that if a section doesn't use any input
machinery, that the user won't pay the price for it.

This was a very complicated change, which should be 100%
invisible to users. It all exists in service of letting
developers write tests that aren't flaky.

Before this change, calling `terminal.type` inside your
run block had a chance of happening so fast that those
keypresses would get dropped because the input machinery
wasn't hooked up yet.

Now, we have hooks in place that notify us when the input
machinery is hooked up, and we block progress in various
places to make sure that by the time the user starts typing
some stuff, we'll be ready for them, even in tests.

Fixes #114