Replace awkward "request block" logic with blocking startup logic
Before, we tried to let various jobs ask nicely if the upstream flow
would wait for them to be ready.

However, it turns out it is easier (and more robust) just to block
progress until these jobs finish connecting to the flows before
continuing.

For a concrete example, here is a test that used to fail:

```
section { }.runUntilSignal {
  onKeyPressed {
    if (key == Keys.Q) signal()
  }
  terminal.press(Keys.Q)
}
```

The reason this would fail is often the `terminal.press(Keys.Q)` line would
get called SO QUICKLY and at JUST SUCH A TIME that its key would get processed
when the upstream flow was ready but the downstream "key processing job" was
not.

Now, what happens is `onKeyPressed` simply blocks, inperceptibly to the human
eye, so that by the time `terminal.press(Keys.Q)` is called, everything is
guaranteed to be hooked up and ready to go.

Fixes #114